refactor: modularize overview components
