+++ frontend/src/components/overview/hooks/useOnboardingStatus.ts
@@ -1,0 +1,12 @@
+import { useState, useEffect } from "react";
+
+export function useOnboardingStatus() {
+  const [isOnboardingDone, setIsOnboardingDone] = useState(true);
+
+  useEffect(() => {
+    const done = localStorage.getItem('onboarding_done_v1') === 'true';
+    setIsOnboardingDone(done);
+  }, []);
+
+  return isOnboardingDone;
+}
