+++ frontend/src/components/overview/lib/fmtCash.ts
@@ -1,0 +1,7 @@
+export function fmtCash(n: number) {
+  return n.toLocaleString(undefined, {
+    style: "currency",
+    currency: "USD",
+    maximumFractionDigits: 0,
+  });
+}
