+++ frontend/src/components/overview/lib/index.ts
@@ -1,0 +1,3 @@
+export { fmtCash } from "./fmtCash";
+export { parseHighlights } from "./parseHighlights";
+export type { HighlightSection } from "./parseHighlights";
